refactor: 인증 관련 정보 관리 로직 분리
